Changing the daemon-launch mechanism
To be perfectly honest, I don't remember why this is changed.  I changed it over a year ago, and have been running that code ever since, so that must mean it's good.... right?
I know I did have problems with launchd occasionally deciding not to launch NanoFibre under Tiger... maybe this was a workaround for that?